Skip TestProcessGroupTermination on {Open,Net}BSD
As it requires github.com/mitchellh/go-ps.
